+++ packages/r4-datamodels/typedoc.config.mjs
@@ -1,0 +1,9 @@
+import { OptionDefaults } from 'typedoc';
+
+/** @type {Partial<import('typedoc').TypeDocOptions>} */
+const config = {
+  extends: '../../typedoc.config.base.mjs',
+  name: '@paq-ts-fhir/r4-datamodels',
+};
+
+export default config;